Move register_provider into ProviderRegistrar
The Container class is long enough, and this centralises all provider handling into the ProviderRegistrar
